+++ src/main.tsx
@@ -1,5 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import '@fontsource-variable/inter'
+import '@fontsource-variable/lexend'
+import '@fontsource-variable/literata'
+import '@fontsource-variable/eb-garamond'
+import './fonts/linguistics-pro.css'
 import './index.css'
 import App from './App'