+++ src/vite-env.d.ts
@@ -1,1 +1,6 @@
 /// <reference types="vite/client" />
+
+declare module '@fontsource-variable/inter';
+declare module '@fontsource-variable/lexend';
+declare module '@fontsource-variable/literata';
+declare module '@fontsource-variable/eb-garamond';